Update script_jmol
Introduced several more variables

#This is the place where we are going to colect and list all technical variables related to the 3d-printers, dimensions, etc. These are important for calculations and we need to put them here in order to be able to change them quickly
 
var JMolStlScale = 10; # Jmol exports 1 A as 10 mm. This means that all dimesnions in Jmol are exported 10x larger. We will work with A, but always multiply the final resuls in mm with this. Scale needs to be devided??? by this. We will keep it here for reference and in case JMol changes this in the future

var MinPrintedObjDimen = 3; # minimum printed dimension of an object (bond, atom, thikness). In the paper is 2.6 but will start with 3 mm

var PrintBedMinDimension = 210 / JMolStlScale; # Introduce the PrintBedMinDimension variable. This is the minimum dimension (X, Y, or Z) of the print bed in mm and corrected for A. For now, it is set manually in the script here, but it might be advisable/required to have a way of setting it from the browser. 

#Thresholds for scale (var PrintScaleFactor) for each rendering scheme if required

#Ball and stick
var MaxPrintScaleFactorBallStick = 1.36 / MinPrintedObjDimen; #soemthing i fail to understand in the Math here, need to check and investigate, needs validation by printing. 1.36 is the C diameter at max 40%vWr that still allows atoms to be separated.
var MinPrintScaleFactorBallStick = (2 * MinPrintedObjDimen) / (2.86 * JMolStlScale); # where 2.86 is the length of a C-C + peptide bonds - based on Math, needs validation by printing

#Sticks - aka backbone -- wil continue here
var MaxPrintScaleFactorBackbone = MaxPrintScaleFactorBallStick; #temporary, needs math
var MinPrintScaleFactorBackbone = 1; #temporary, from the paper, needs math